Fixed seen bug
Fixed a bug where the seen set was not being utilised due to using Longs instead of numbers. Longs cannot be natively used in sets as they were, so I've migrated the cars to use 48 instead of 64 bits, as that can safely be fit inside a number. I have also changed the seen set to use numbers.