Fixed production bug caused by StrictMode
In development I was using StrictMode, and as I was new to React this caused me to find a hacky solution to a problem that was caused by this, not knowing an alternative method. Now, I have fixed the core problem, and deleted the hacky solution. Production should be fixed.